Size type changes for the posix implementation

+++ codebase/src/cpp/syscommon/src/Platform.cpp
@@ -1476,7 +1476,7 @@ std::string Platform::toAnsiString( const char* string )
 	return std::string( string );
 }
 
-std::string Platform::toAnsiString( const char* string, size_t length )
+std::string Platform::toAnsiString( const char* string, int length )
 {
 	return std::string( string, length );
 }
@@ -1486,7 +1486,7 @@ std::string Platform::toAnsiString( const wchar_t* string )
 	return toAnsiString( string, ::wcslen(string) );
 }
 
-std::string Platform::toAnsiString( const wchar_t* string, size_t length )
+std::string Platform::toAnsiString( const wchar_t* string, int length )
 {
 	std::string result;
 	return result;
@@ -1502,10 +1502,10 @@ int Platform::toUnicodeChars( const char* string,
 
 std::wstring Platform::toUnicodeString( const char* string )
 {
-	return Platform::toUnicodeString( string, strlen(string) );
+	return Platform::toUnicodeString( string, (int)strlen(string) );
 }
 
-std::wstring Platform::toUnicodeString( const char* string, size_t length )
+std::wstring Platform::toUnicodeString( const char* string, int length )
 {
 	std::wstring result;
 	return result;
@@ -1516,7 +1516,7 @@ std::wstring Platform::toUnicodeString( const wchar_t* string )
 	return std::wstring( string );
 }
 
-std::wstring Platform::toUnicodeString( const wchar_t* string, size_t length )
+std::wstring Platform::toUnicodeString( const wchar_t* string, int length )
 {
 	return std::wstring( string, length );
 }
@@ -1548,7 +1548,7 @@ String Platform::toPlatformString( const char* string )
 #endif
 }
 
-String Platform::toPlatformString( const char* string, size_t length )
+String Platform::toPlatformString( const char* string, int length )
 {
 #ifndef UNICODE
 	return Platform::toAnsiString( string, length );
